requirements.txt
streamlit==1.43.1
python-dotenv==1.0.1
google-generativeai==0.8.4
Pillow==11.1.0
